# Conflicts:
#	src/main/java/com/fivepigs/app/dao/SoftwareDao.java
#	src/main/webapp/images/assets/uploads/avatars/user_9_9254e656c3464999a59ed1f4a285a57b.jpg
#	target/classes/com/fivepigs/app/config/Db.class
#	target/classes/com/fivepigs/app/dao/SoftwareDao.class
#	target/classes/com/fivepigs/app/dao/UserDao.class
#	target/classes/com/fivepigs/app/model/ApprovalProcess.class
#	target/classes/com/fivepigs/app/model/Category.class
#	target/classes/com/fivepigs/app/model/ReviewerProcess.class
#	target/classes/com/fivepigs/app/model/Software.class
#	target/classes/com/fivepigs/app/model/SoftwareVersion.class
#	target/classes/com/fivepigs/app/web/DashboardServlet.class

+++ src/main/java/com/fivepigs/app/web/customer/ProductServlet.java
@@ -112,7 +112,7 @@ public class ProductServlet extends HttpServlet {
         return detail.getCreatedAt();
     }
 
-    private Integer resolveFileSize(Software detail) {
+    private Long resolveFileSize(Software detail) {
         SoftwareVersion version = detail.getSoftwareVersion();
         return version == null ? null : version.getFileSize();
     }
@@ -144,7 +144,7 @@ public class ProductServlet extends HttpServlet {
         return dateTime.format(DATE_FORMAT);
     }
 
-    private String formatFileSize(Integer fileSize) {
+    private String formatFileSize(Long fileSize) {
         if (fileSize == null || fileSize <= 0) {
             return "Updating";
         }
